server/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
customtkinter==5.2.2
daphne==4.1.2
darkdetect==0.8.0
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.2.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
kombu==5.4.2
Markdown==3.7
marshmallow==3.23.1
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pi==0.1.2
pillow==11.0.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==24.3.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
qrcode==8.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
v==1
vine==5.1.0
wcwidth==0.2.13
zipp==3.21.0
zope.interface==7.2
psycopg2>=2.9
python-dotenv==1.0.1